requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.5
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
psycopg2==2.9.11
pycparser==2.23
PyJWT==2.11.0
python-dotenv==1.2.1
referencing==0.36.2
rpds-py==0.27.1
typing_extensions==4.15.0
Werkzeug==3.1.6
zipp==3.23.0
